готовый кроме умножения 12

+++ calculator.java
@@ -5,11 +5,11 @@ public class calculator {
         Scanner scanner = new Scanner(System.in);
         System.out.println("=== Калькулятор ===");
         System.out.print("Введите первое число: ");
-        double a = scanner.nextInt();
+        double a = scanner.nextDouble();
         System.out.print("Введите операцию (+, -, *, /): ");
         String op = scanner.next();
         System.out.print("Введите второе число: ");
-        double b = scanner.nextInt();
+        double b = scanner.nextDouble();
         double result;
         if (op.equals("+")) {
             result = add.add1(a, b);
